+++ create-docs.ps1
@@ -1,0 +1,7 @@
+# Reqiuires you to add docfx to system path
+
+docfx metadata
+docfx build
+
+# Requires wkhtmltopdf to be added to system path
+docfx pdf